+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## 4.0 GUI and Plugin Refactoring
 
-- 2023-01-16 4.0.34
+- 2023-01-25 4.0.35
+    - RamanModeFeature 
+        - prompt for confirmation
+        - release restriction on LaserControlFeature when invisible
+- 2023-01-24 4.0.34
     - Refactored RamanModeFeature, VCRControls and TakeOneFeature around TakeOneRequest
     - moved to ctl
         - BatchCollection